+++ components/blitz/src/omero/gateway/model/SearchDataContext.java
@@ -20,7 +20,7 @@
  *
  *------------------------------------------------------------------------------
  */
-package omero.gateway;
+package omero.gateway.model;
