+++ components/blitz/src/omero/gateway/model/SecurityContext.java
@@ -21,7 +21,7 @@
  *
  *------------------------------------------------------------------------------
  */
-package omero.gateway;
+package omero.gateway.model;
